requirements.txt
flask
pymongo
redis
python-dotenv
requests
beautifulsoup4
bcrypt
pyjwt
gunicorn
pydantic[email]
prometheus-client
